+++ src/Field.php
@@ -268,6 +268,7 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
                 $newBlockType->description = $blockType['description'];
                 $newBlockType->minBlocks = (int)$blockType['minBlocks'];
                 $newBlockType->maxBlocks = (int)$blockType['maxBlocks'];
+                $newBlockType->minSiblingBlocks = (int)$blockType['minSiblingBlocks'];
                 $newBlockType->maxSiblingBlocks = (int)$blockType['maxSiblingBlocks'];
                 $newBlockType->minChildBlocks = (int)$blockType['minChildBlocks'];
                 $newBlockType->maxChildBlocks = (int)$blockType['maxChildBlocks'];