+++ src/assets/InputAsset.php
@@ -183,6 +183,7 @@ class InputAsset extends FieldAsset
                 'description' => $blockType->description,
                 'minBlocks' => $blockType->minBlocks,
                 'maxBlocks' => $blockType->maxBlocks,
+                'minSiblingBlocks' => $blockType->minSiblingBlocks,
                 'maxSiblingBlocks' => $blockType->maxSiblingBlocks,
                 'minChildBlocks' => $blockType->minChildBlocks,
                 'maxChildBlocks' => $blockType->maxChildBlocks,