+++ src/models/BlockType.php
@@ -74,6 +74,12 @@ class BlockType extends Model implements GqlInlineFragmentInterface
     public ?int $maxBlocks = null;
 
     /**
+     * @var int|null The minimum number of blocks of this type allowed under one parent block.
+     * @since 3.3.0
+     */
+    public ?int $minSiblingBlocks = null;
+
+    /**
      * @var int|null The maximum number of blocks of this type allowed under one parent block.
      * @since 2.8.0
      */
@@ -173,7 +179,7 @@ class BlockType extends Model implements GqlInlineFragmentInterface
     {
         return [
             [['id', 'fieldId', 'sortOrder'], 'number', 'integerOnly' => true],
-            [['minBlocks', 'maxBlocks', 'minChildBlocks', 'maxChildBlocks'], 'integer', 'min' => 0],
+            [['minBlocks', 'maxBlocks', 'minChildBlocks', 'maxChildBlocks', 'minSiblingBlocks'], 'integer', 'min' => 0],
         ];
     }
 
@@ -263,6 +269,7 @@ class BlockType extends Model implements GqlInlineFragmentInterface
             'maxBlocks' => (int)$this->maxBlocks,
             'minChildBlocks' => (int)$this->minChildBlocks,
             'maxChildBlocks' => (int)$this->maxChildBlocks,
+            'minSiblingBlocks' => (int)$this->minSiblingBlocks,
             'maxSiblingBlocks' => (int)$this->maxSiblingBlocks,
             'name' => $this->name,
             'sortOrder' => (int)$this->sortOrder,